requirements.txt
asgiref==3.8.1
Django==5.1.7
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.15.2
pillow==11.1.0
psycopg==3.2.5
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
